// src/data/baseline.generated.ts
export const GENERATED_AT = '2026-08-27'
// ...
export const COMPAT_KEYS: Record<string, { support: Record<string, string> }> = {
  "api.AbortController.AbortController": {
    "support": {
      "chrome": "66",
      "chrome_android": "66",
      "edge": "16",
      "firefox": "57",
      "firefox_android": "57",
      "safari": "12.1",
      "safari_ios": "12.2"
    }
  },
// ...
}